requirements.txt
gspread
oauth2client
pandas
Pillow
python-dotenv
anthropic
google-generativeai
emoji
feedparser 
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2